Assume A entered in to room-->bottom sheet-->join room-->click on 3 dots-->view members -->click on back button two times -->then alert box is not displaying.
Assume A create new room-->system message is displayed  and one send message -->then click back button of the device-->dialog box is not displaying to stay subscribe or unsubscribe.

Assume A and B-->entered into same room -->A click back button of the device-->then system message is not displaying in B's end  but when you click back arrow then system message is displaying.

 Assume A entered public room and click on START-->list of rooms is displaying-->go to BG and comes to FG-->then app is not retaining the state